+++ demo-app/jest/css.js
@@ -3,6 +3,9 @@
  *
  * @format
  * @flow strict
+ * @emails oncall+internationalization
  */
 
+// Empty CSS mock used with jest so that we can import `.css` files like this:
+//   import './css/Example.css';
 module.exports = ({}: {});