+++ packages/babel-plugin-fbt/src/bin/translate.bin.js
@@ -1,0 +1,12 @@
+#!/usr/bin/env node
+/**
+ * Copyright 2004-present Facebook. All Rights Reserved.
+ *
+ * @format
+ * @noflow
+ * @emails oncall+internationalization
+ */
+
+'use strict';
+
+require('./translate');