+++ packages/fb-tiger-hash/src/index.js
@@ -2,7 +2,9 @@
  * Copyright 2004-present Facebook. All Rights Reserved.
  *
  * @emails oncall+internationalization
+ *
  * @format
+ * @flow strict
  */
 
 'use strict';
@@ -13,6 +15,6 @@ const hashPhrases = require('./hashPhrases');
 module.exports = {
   __esModule: true,
   default: Tiger,
-  Tiger: Tiger,
-  hashPhrases: hashPhrases,
+  Tiger,
+  hashPhrases,
 };