+++ runtime/shared/__tests__/fbt-test.js
@@ -40,19 +40,13 @@ let intlNumUtils;
 describe('fbt', () => {
   beforeEach(() => {
     jest.resetModules();
-
     intlNumUtils = jest.requireActual<intlNumUtilsType>('intlNumUtils');
     fbtRuntime = jest.requireActual('fbt');
     fbt = require('fbt');
     domContainer = document.createElement('div');
-
-    // Use a locale that has FEW.
-    FbtNumberType.getVariation = jest.requireActual(
-      'IntlCLDRNumberType19',
-    ).getVariation;
   });
 
-  it('should memoize new strings', function() {
+  it('should memoize new strings', function () {
     expect(fbtRuntime._getCachedFbt('sample string')).toEqual(undefined);
 
     expect(fbtRuntime._('sample string')).toEqual(
@@ -60,17 +54,17 @@ describe('fbt', () => {
     );
   });
 
-  it('should trivially handle tokenless strings', function() {
+  it('should trivially handle tokenless strings', function () {
     expect(fbt('without tokens', 'test')).toEqual('without tokens');
   });
 
-  it('should handle common strings', function() {
+  it('should handle common strings', function () {
     expect(fbt.c('Accept')).toEqual(
       fbt('Accept', 'Button/Link: Accept conditions'),
     );
   });
 
-  it('should replace tokens with named values', function() {
+  it('should replace tokens with named values', function () {
     expect(
       fbt('with token ' + fbt.param('token', 'A') + ' here', 'test'),
     ).toEqual('with token A here');
@@ -86,16 +80,16 @@ describe('fbt', () => {
     ).toEqual('with tokens A and B');
   });
 
-  it('should remove punctuation when a value ends with it', function() {
-    expect(
-      fbt('They said ' + fbt.param('quote', '"Hi!"') + '.', 'test'),
-    ).toEqual('They said "Hi!"');
-    expect(fbt('They said ' + fbt.param('quote', 'bye') + '.', 'test')).toEqual(
-      'They said bye.',
+  it('should remove punctuation when a value ends with it', function () {
+    expect(fbt('Play ' + fbt.param('game', 'Chess!') + '!', 'test')).toEqual(
+      'Play Chess!',
     );
+    expect(
+      fbt("What's on your mind " + fbt.param('name', 'T.J.') + '?', 'test'),
+    ).toEqual("What's on your mind T.J.?");
   });
 
-  it('should allow values that look like token patterns', function() {
+  it('should allow values that look like token patterns', function () {
     expect(
       fbt(
         'with tokens ' +
@@ -108,7 +102,7 @@ describe('fbt', () => {
     ).toEqual('with tokens {tokenB} and B');
   });
 
-  it('should support objects as token values', function() {
+  it('should support objects as token values', function () {
     // We expect that this returns an opaque React fragment instead of an array.
     // We use this to preserve identity of nested React elements.
     const argument = <div />;
@@ -117,13 +111,13 @@ describe('fbt', () => {
       'test',
     );
     const items = [];
-    React.Children.forEach(fragment, function(item) {
+    React.Children.forEach(fragment, function (item) {
       items.push(item);
     });
     expect(items).toEqual(['with token ', argument, ' here']);
   });
 
-  it('should render empty string for null values', function() {
+  it('should render empty string for null values', function () {
     expect(fbt(fbt.param('null_value', null), 'test')).toEqual('');
   });
 
@@ -160,7 +154,7 @@ describe('fbt', () => {
     }
   }
 
-  it('should use wildcard defaults', function() {
+  it('should use wildcard defaults', function () {
     expect(
       fbt(
         'with something like ' +
@@ -171,7 +165,7 @@ describe('fbt', () => {
     ).toEqual('with something like 42 wildcards');
   });
 
-  it('should format numeric value', function() {
+  it('should format numeric value', function () {
     expect(
       fbt(
         'A total amount is ' + fbt.param('count', 10000, {number: true}),
@@ -180,7 +174,7 @@ describe('fbt', () => {
     ).toEqual('A total amount is 10,000');
   });
 
-  it('should keep literal value as is', function() {
+  it('should keep literal value as is', function () {
     expect(
       fbt('A total amount is ' + fbt.param('count', 10000), 'Test string'),
     ).toEqual('A total amount is 10000');
@@ -188,30 +182,30 @@ describe('fbt', () => {
 
   function renderAndExtractChildDivs(component) {
     const node = ReactDOM.findDOMNode(ReactDOM.render(component, domContainer));
-    // TODO T21716504: flow thinks ReactDOM.findDOMNode returns Text...
+    // flow thinks ReactDOM.findDOMNode can return a type of Text...
     invariant(node instanceof Element, 'Expected node to be an Element');
-    const resultingElements = node && node.getElementsByTagName('div');
+    const resultingElements = node.getElementsByTagName('div');
     return Array.prototype.slice.call(resultingElements, 0);
   }
 
-  it('should not warn when unkeyed React components are params', function() {
+  it('should not warn when unkeyed React components are params', function () {
     jest.spyOn(console, 'warn').mockImplementation(() => {});
     const nodes = renderAndExtractChildDivs(
-      <TestComponent value="A" childA={<div />} childB={<div />} />,
+      <TestComponent childA={<div />} childB={<div />} value="A" />,
     );
     expect(nodes.length).toBe(2);
-    /* $FlowFixMe(>=0.99.0 site=www) This comment suppresses an error found
-     * when Flow v0.99 was deployed. To see the error delete this comment and
-     * run Flow. */
+    /* $FlowFixMe[prop-missing] (>=0.99.0 site=www) This comment suppresses an
+     * error found when Flow v0.99 was deployed. To see the error delete this
+     * comment and run Flow. */
     expect(console.warn.mock.calls.length).toBe(0);
   });
 
   function expectChildSetsToRetainIdentity(setA, setB) {
     const nodesA = renderAndExtractChildDivs(
-      <TestComponent value="A" childA={setA} childB={setB} />,
+      <TestComponent childA={setA} childB={setB} value="A" />,
     );
     const nodesB = renderAndExtractChildDivs(
-      <TestComponent value="B" childA={setA} childB={setB} />,
+      <TestComponent childA={setA} childB={setB} value="B" />,
     );
 
     expect(nodesA.length).toBe(2);
@@ -222,7 +216,7 @@ describe('fbt', () => {
     expect(nodesA[1]).toBe(nodesB[0]);
   }
 
-  it('should retain React identity when sentence order changes', function() {
+  it('should retain React identity when sentence order changes', function () {
     expectChildSetsToRetainIdentity(<div key="A" />, <div key="B" />);
   });
 
@@ -240,13 +234,13 @@ describe('fbt', () => {
   //   );
   // });
 
-  it('should replace QuickTranslation strings', function() {
+  it('should replace QuickTranslation strings', function () {
     expect(
       fbtRuntime._(['This is a QT string', '8b0c31a270a324f26d2417a358106611']),
     ).toEqual('override');
   });
 
-  it('should replace QuickTranslation strings with params', function() {
+  it('should replace QuickTranslation strings with params', function () {
     expect(
       fbtRuntime._(
         ['Just a {param}', 'fakeHash3'],
@@ -255,7 +249,7 @@ describe('fbt', () => {
     ).toEqual('Override a substitute');
   });
 
-  it('should replace QuickTranslation "trees"', function() {
+  it('should replace QuickTranslation "trees"', function () {
     const runtimeArg = {
       s: ['This is a QT with a {param}', 'fakeHash1'],
       p: ['These are QTs with a {param}', 'fakeHash2'],
@@ -275,32 +269,7 @@ describe('fbt', () => {
     ).toEqual('These are overrides and a test');
   });
 
-  it('should handle variated numbers', function() {
-    const previousNumberTypeGetter = FbtNumberType.getVariation;
-    FbtNumberType.getVariation = jest.requireActual(
-      'IntlCLDRNumberType31',
-    ).getVariation;
-    const numToType = {
-      '21': IntlVariations.NUMBER_ONE,
-      '22': IntlVariations.NUMBER_TWO,
-      '103': IntlVariations.NUMBER_FEW,
-      '1000000': IntlVariations.NUMBER_MANY,
-      '15': IntlVariations.NUMBER_OTHER,
-    };
-    for (const n in numToType) {
-      const type = numToType[n];
-      const displayNumber = intlNumUtils.formatNumberWithThousandDelimiters(
-        parseFloat(n),
-      );
-      expect(fbtRuntime._param('num', parseInt(n, 10), [0])).toEqual([
-        [type, '*'],
-        {num: displayNumber},
-      ]);
-    }
-    FbtNumberType.getVariation = previousNumberTypeGetter;
-  });
-
-  it('should create a tuple for fbt.subject if valid', function() {
+  it('should create a tuple for fbt.subject if valid', function () {
     expect(fbtRuntime._subject(IntlVariations.GENDER_MALE)).toEqual([
       [IntlVariations.GENDER_MALE, '*'],
       null,
@@ -308,13 +277,13 @@ describe('fbt', () => {
     expect(() => fbtRuntime._subject(0)).toThrow('Invalid gender provided');
   });
 
-  it('should leave non-QuickTranslation strings alone', function() {
+  it('should leave non-QuickTranslation strings alone', function () {
     expect(
       fbtRuntime._(["This isn't", '8b0c31a270a324f26d2417a358106612']),
     ).toEqual("This isn't");
   });
 
-  it('should access table with multiple tokens containing subject', function() {
+  it('should access table with multiple tokens containing subject', function () {
     expect(
       fbt(
         'Invited by ' + fbt.plural('friend', 1, {showCount: 'yes'}) + '.',
@@ -322,83 +291,6 @@ describe('fbt', () => {
         {subject: IntlVariations.GENDER_UNKNOWN},
       ),
     ).toEqual('Invited by 1 friend.');
-  });
-
-  it('should access table with fallback logic', function() {
-    const FbtHooks = require('FbtHooks');
-    let genderMock;
-    // $FlowFixMe We need to mock this method
-    FbtHooks.getViewerContext = jest.fn(() => ({
-      GENDER: genderMock,
-    }));
-
-    const table = {
-      __vcg: 1, // viewer-context gender
-      '*': {},
-    };
-    table['*']['A'] = {'*': 'A,UNKNOWN,OTHER {name} has {num}'};
-    table['*']['A'][ONE] = 'A,UNKNOWN,ONE {name} has {num}';
-    table['*']['A'][FEW] = 'A,UNKNOWN,FEW {name} has {num}';
-    table['*']['B'] = {'*': 'B,UNKNOWN,OTHER {name} has {num}'};
-    table['*']['B'][ONE] = 'B,UNKNOWN,ONE {name} has {num}';
-    table['*']['B'][FEW] = 'B,UNKNOWN,FEW {name} has {num}';
-    table[MALE] = {A: {'*': 'A,MALE,OTHER {name} has {num}'}};
-    table[MALE]['A'][ONE] = 'A,MALE,ONE {name} has {num}';
-    // table['*'][male]['A'][FEW] = fallback to other ^^^
-    // table['*'][male]['B'] = fallback to unknown gender ^^^
-    table[FEMALE] = {B: {'*': 'B,FEMALE,OTHER {name} has {num}'}};
-    table[FEMALE]['B'][FEW] = 'B,FEMALE,FEW {name} has {num}';
-    // table[female]['B'][ONE] = fallback to other ^^^
-    // table[female]['A'] = fallback to unknown gender ^^^
-
-    const few = fbtRuntime._param('num', 10, [0] /*Variations.NUMBER*/);
-    const other = fbtRuntime._param('num', 20, [0]);
-    const one = fbtRuntime._param('num', 1, [0]);
-    const A = fbtRuntime._enum('A', {A: 'A', B: 'B'});
-    const B = fbtRuntime._enum('B', {A: 'A', B: 'B'});
-    const name = fbtRuntime._param('name', 'Bob');
-
-    // GENDER UNKNOWN
-    genderMock = IntlVariations.GENDER_UNKNOWN;
-    let tests = [
-      {arg: [A, few, name], expected: 'A,UNKNOWN,FEW Bob has 10'},
-      {arg: [A, one, name], expected: 'A,UNKNOWN,ONE Bob has 1'},
-      {arg: [A, other, name], expected: 'A,UNKNOWN,OTHER Bob has 20'},
-      {arg: [B, few, name], expected: 'B,UNKNOWN,FEW Bob has 10'},
-      {arg: [B, one, name], expected: 'B,UNKNOWN,ONE Bob has 1'},
-      {arg: [B, other, name], expected: 'B,UNKNOWN,OTHER Bob has 20'},
-    ];
-    const runTest = function(test) {
-      try {
-        expect(fbtRuntime._(table, test.arg)).toBe(test.expected);
-      } catch (error) {
-        error.message += `\ntest.expected="${test.expected}"`;
-        throw error;
-      }
-    };
-    tests.forEach(runTest);
-
-    genderMock = IntlVariations.GENDER_MALE;
-    tests = [
-      {arg: [A, few, name], expected: 'A,MALE,OTHER Bob has 10'},
-      {arg: [A, one, name], expected: 'A,MALE,ONE Bob has 1'},
-      {arg: [A, other, name], expected: 'A,MALE,OTHER Bob has 20'},
-      {arg: [B, few, name], expected: 'B,UNKNOWN,FEW Bob has 10'},
-      {arg: [B, one, name], expected: 'B,UNKNOWN,ONE Bob has 1'},
-      {arg: [B, other, name], expected: 'B,UNKNOWN,OTHER Bob has 20'},
-    ];
-    tests.forEach(runTest);
-
-    genderMock = IntlVariations.GENDER_FEMALE;
-    tests = [
-      {arg: [A, few, name], expected: 'A,UNKNOWN,FEW Bob has 10'},
-      {arg: [A, one, name], expected: 'A,UNKNOWN,ONE Bob has 1'},
-      {arg: [A, other, name], expected: 'A,UNKNOWN,OTHER Bob has 20'},
-      {arg: [B, few, name], expected: 'B,FEMALE,FEW Bob has 10'},
-      {arg: [B, one, name], expected: 'B,FEMALE,OTHER Bob has 1'},
-      {arg: [B, other, name], expected: 'B,FEMALE,OTHER Bob has 20'},
-    ];
-    tests.forEach(runTest);
   });
 
   it('should have a return type compatible with the public Fbt type', () => {
@@ -435,13 +327,13 @@ describe('fbt', () => {
 
       expect(
         fbt(
-          /* $FlowFixMe(>=0.121.0 site=www) This comment suppresses an error
-           * found when Flow v0.121 was deployed. To see the error delete this
-           * comment and run Flow. */
+          /* $FlowFixMe[incompatible-call] (>=0.121.0 site=www) This comment
+           * suppresses an error found when Flow v0.121 was deployed. To see
+           * the error delete this comment and run Flow. */
           fbt.param('hello', fbtParams[0]) +
-            /* $FlowFixMe(>=0.121.0 site=www) This comment suppresses an error
-             * found when Flow v0.121 was deployed. To see the error delete
-             * this comment and run Flow. */
+            /* $FlowFixMe[incompatible-call] (>=0.121.0 site=www) This comment
+             * suppresses an error found when Flow v0.121 was deployed. To see
+             * the error delete this comment and run Flow. */
             fbt.param('world', fbtParams[1]),
           'desc',
         ),