+++ runtime/shared/intlList.js
@@ -41,13 +41,13 @@ const DELIMITERS = {
   SEMICOLON: 'SEMICOLON',
 };
 
-const intlList = function<TItem: React.Node>(
+const intlList = function <TItem: React.Node>(
   items: $ReadOnlyArray<TItem>,
   conjunction: ?$Keys<typeof CONJUNCTIONS>,
   delimiter: ?$Keys<typeof DELIMITERS>,
 ): TItem | Fbt {
   if (__DEV__) {
-    items.forEach(function(item) {
+    items.forEach(function (item) {
       invariant(
         typeof item === 'string' || React.isValidElement(item),
         'Must provide a string or ReactComponent to intlList.',
@@ -179,12 +179,15 @@ function _getConjunction(
 intlList.DELIMITERS = DELIMITERS;
 intlList.CONJUNCTIONS = CONJUNCTIONS;
 
-module.exports = (intlList: (<TItem: React.Node>(
-  items: $ReadOnlyArray<TItem>,
-  conjunction: ?$Keys<typeof CONJUNCTIONS>,
-  delimiter: ?$Keys<typeof DELIMITERS>,
-) => TItem | Fbt) & {
+type IntlListExport = {|
+  <TItem: React.Node>(
+    items: $ReadOnlyArray<TItem>,
+    conjunction: ?$Keys<typeof CONJUNCTIONS>,
+    delimiter: ?$Keys<typeof DELIMITERS>,
+  ): TItem | Fbt,
+
   DELIMITERS: typeof DELIMITERS,
   CONJUNCTIONS: typeof CONJUNCTIONS,
-  ...
-});
+|};
+
+module.exports = (intlList: IntlListExport);